Bug fix on release branch: #2039222 Header doesn't work in IE
This is due to the fact that IE only supports :hover on 
<a> tags, and we are using them on divs also. Firefox and
most other browsers support this standard behavior, so it
was not discovered earlier.

The fix is to add csshover.htc which is a script that only
runs in IE and it adds javascript actions to all of the
:hover elements to fix IE's default behavior.

The script is LGPL licensed, so it is safe to use here.