This is a dummy commit -- no real change.
I realized that the previous commit was marked as merging
revisions 754:771, but in fact it actually merged 754:783.

I thought I would have to merge twice due to the branch 
being renamed at revision 771, but subversion took care
of that for me.

Next week's merge will start at 784 and go on from there.