Fixing a few things I found:
1) The Notify.php was trying to use $_SERVER['HOST_NAME'], which is not defined
for command line scripts. My solution is to modify the installer to store the
host URL in the install.conf. All of the command line scripts in OpenFISMA will
be able to read the value from there.

2) While working on #1, I found that the base.sql contained a lot of corrupted
data in the base.sql file, so I cleaned out that data. (Random bytes of data
in the middle of large blocks of text.)

3) I also modified testmigrate.pl to dump the schema in a slighly different
format, so that base.sql is a little easier to read.

4) Added blank migration called #19 which is only used to regenerate the base.sql
using the new testmigrate.pl script.