Fixing bug #2149199
The fix is to set the positioning attribute to "relative" in IE7 only. 

A conditional comment is added to the layout.phtml which provides an ie7hacks.css file only for IE7 browsers. All future IE7 hacks should go into this file.

The same pattern should be used if we need to work around quirks in other versions of IE.