Adding audit logging to the injection plugin.
After investigation, I found out that audit loggging was not being done in the Poam class, but was actually be done in several different places throughout the application. This is a perfect example of why we want to consolidate business logic in the models as much as possible. At the same time, this also fixes a bug where findings uploaded in a spreadsheet did not have audit events created.