Bug fix on trunk #2166876 FIPS 199 Compliance Failure
Confidentiality is different from Integrity and Availability with the increased value "NA"
the HTML RAF and  PDF RAF need to be exactly the same

we have redefined the method for calculate the security categorization, and resolved the changings on the bug report one by one, except one representation style of PDF, i.e. the shade highlight cell. The PDF generation library does not support shading an individual cell.