Bug fix on trunk #2412141 Privilege exceptions are not handled correctly.
Firstly, it's caused by the switch of throw exceptions. When it's turned on, usually in development mode, any exception will not be handled by the handler. Instead it goes to the end user.

Secondly, as we used actionStack for a serial of actions and privilege requirement failure results in an exception, a serial of exceptions occurs even after error handler is involved in. Then the ZF think exceptions happens in error handler and it disable error handling but directly throw it to browsers.

To solve these, we have:
1. Make the throw exception switch configurable in app.ini
2. Clear the action stack when handling any exception
Besides of this, we follow the suggestions on the SF.net.