Bug fix on trunk #2390549 Replace the status OPEN with the status DRAFT
In previous commit r1259, the schema was upgraded without altering the existing data( status=="OPEN" ). The result of the upgrade is the former record of OPEN turns out to be empty(""). 

Because of this r1259 is flawed and MUST NOT be tagged or release. Then it has to be updated the script of 44.up.sql&44.dn.sql instead of add a 45.xx.sql. 
The change makes existing OPEN poams to "DRAFT".