Bug fix on trunk #2414983 Mitigation strategy history is not displayed in DRAFT status
root cause:
To simplify the logic, the history of mitigation's evaluation is deleted after a revising of it. Then there is no any history information of evaluation when the poam revisiting DRAFT status from EN. However the history perserves when a poam is back to DRAFT from MSA status, when the SSO denies the mitigation strategy.

It's suggested to clean up(Delete) the MS evaluation history whenever a 'Submit MS' is commit. However to keep the maximum compatibility backward, history information perserves when a poam moves from MSA to DRAFT. The last evaluation tells the difference(APPROVE='EN=>DRAFT', DENIED='MSA=>DRAFT')