Bug fix on release branch 2.2 #2414983 Mitigation strategy history is not displayed in DRAFT status
root casue: We completed partial of the work to specially treat the MSA denial in terms of evaluation history. If the last evaluation is 'DENIED', the history is kept however no guardian is there for DRAFT status. Then a new round of evaluation starts at the improper place.

Guard the DRAFT status for the mitigation strategy history.