Bug fix on release branch 2.2 Installer can be executed by any user
All the actions of the install controller should be and was disabled except the default one. In case the URL ,say 'install', goes without the trailing ‘/’, the ZF MVC will ,by default, come to the controller name and ‘index’ action of it. That escapes the prevention of the route, which is matched by regex ‘^install/’. However, it cannot go futher to install anything in fact.

Change the regex from ‘^install/’ to ‘install’ to prevent the whole install controller.