Bug fix on trunk #2418149 When login expires, clicking on a link creates an error
not check whether the user login or not, when the error controller handle a exception.

When an exception is handled, it's checked whether the user has login or not(expired).
Yes: show the error message with header and footer.
No: show the error message on login page.

Also, an error message is displayed on the login page, when a user access a path which doesn't exist before he logins.