Bug fix on release 2.2 #2462650 RAF Generation
RAFs generation page has submit a form whose action is pointed at a panel related url. It stacks the header and footer in a layout.
While outputting the TGZ file, the header and footer contents(HTML tags) come along although the layout is disabled. The stack FILO
feature contributes to this also. 

Why to use /panel/sub/report/rafs instead of /report/rafs? It's because when there is no findings belonging to the system, a blank 
page with a bare error message present due to the absence of layout, which is introduced by /panel/header

to fix:
1. Restore the action of the form to /report/rafs.
2. Use _forward(...) with additional parameter to fix the bare error message page issue.