Feature request on trunk #2505135  Improve the Search feature - Customize the column display
Users are allowed to show & hide the columns as they see fit. 
A user does this by clicking on the "gear" icon in the header. 
A window will popup which will allow the users to select which columns to view.

The user's preference for which columns to show and which columns to hide is saved by the system; 
on subsequent logins, the system will use the same display preferences. 

The cookie is not a persistent cookies.