#2529181  	 RAF - threat likelihood calculation is incorrect
Modify the logic to use a lookup table instead. Previously we were trying to reduce dependencies on constants and get this information directly from the DBMS, but that code was really complicated to make it work. For constants and calculations that are defined in federal standards (this one is defined in NIST SP 800-30) we should put the constants and calculations directly into our code in order to make the code more readable.

I'm also including a unit test for this fix.