Bug fix on release 2.2 #2569337 Sws_Exception
There is a reference to an invalid class called Sws_Exception in one of the initialize plugins.

I removed this Exceptions cause the other exception will be threw before.
We needn't depend this exception to judge whether the database is existed or whether the configuration is right.

to see Config::readSysConfig line 308