Bug fix on release 2.2 #2568917 Unit tests & CLI Scripts do not bootstrap correctly
Currently, the application uses a bootstrap plugin.

However, the CLI scripts and the unit tests bootstrap using the
bootstrap.php file.

Update the CLI scripts and unit tests to use their corresponding bootstrap
plugins.

to fix:
1. Rename Plugin_Initialize_Webapp to Plugin_Bootstrap_WebApp. Rename the
others in the same way.
2. Update the unit tests and notifier scripts to use the bootstrap plugin.
Currently they just import the bootstrap.php file.

PS: Though I Update the notifier scripts to use the bootstrap plugin,
there are still some errors in it, so I cannot test it. Refer to "Notification events" doesn't work. - ID: 2571819
As this is fixed on trunk by Ryan it is expected to work on next release after merging.