Feature request on YUI branch: 2646656 Migration to YUI
Now the column preferences can be saved into the users table in the search_column_pref field.

I changed the type of this field from varchar to integer to create a proper bitmask.

Still cannot figure out how to style the column preferences panel itself. Right now it's just a jumble of buttons.