Feature Request on trunk #2646656 Migration to YUI
The spreadsheet was broken because in a previous merge this didn't get copied over. I am copying the relevant files from the 2.2 release branch and cut and pasted the new functions into the Finding Controller. I did some quick testing and it is working okay, but I need to test some more.