Feature Request on trunk #2646656 Migration to YUI
Fixing the IE6, IE7 bug where the mitigation tab doesn't work in MSA status and the artifacts tab doesn't work when in EA status.

The reason is that MSA and EA use a nested <form> tag, and when IE loads this into an innerHtml attribute, it fails.

The fix is a bit of a hack: remove the nested form. In the controller, intercept the preDispatch() and forward approval forms to the corresponding controller action. The denial javascripts have to add an attribute in order to let the controller know it's an approval action.