Feature Request on trunk #2646656 Migration to YUI
Fixing the issue with not being able to display the evidence upload dialog. This is similar to the previous commit where a nested form cannot be inserted using innerHtml. This one a little more complicated because JQuery doesn't put the innerHtml content inside the original form, so I had to switch from the Jquery dialog to a YUI dialog.