Feature Request on trunk #2646656 Migration to YUI
I noticed that if the finding source contains an ampersand (&), then the YUI data table truncates the value. So the quick fix for our client is to use htmlenties() on the source name when rendering the table data.

In the future this will need to be spec'ed out and done in a broader, more consistent way.