Bug fix on branch release 2.3 #2776402 Create file upload table
1) Modify the spreadsheet upload and scan report upload to add a timestamp
to the name of the uploaded file (like the evidence uploads). The timestamp
should be in the following format: YYYYMMDDHHmmSS. (Year, Month Day, Hours,
Minutes, Seconds).

By specifying time down to the precision of seconds, the probability of two
files having the same file name (and thus overwriting one of the files) is
very small.

2) Create a new table called "uploads" which contains the following data:

id: primary key
upload_ts: timestamps when the file was uploaded
user_id: ID of the user who uploaded the file
filename: name of the uploaded file

3) Modify the "poams" table to include a foreign key to the "uploads" table
and call it "upload_id". "upload_id" is a null able column.

4) Modify the spreadsheet upload and scan report upload to create a row in
the "uploads" table for each file that is uploaded successfully. If a file
upload fails, then a row should not be created. (Use DBMS transactions to
ensure atomicity.)

Also, for each finding that is created, insert the upload_id with a value
corresponding to the file it came from.

5) Modify the finding data tab. Where it says "Source", if the upload_id is
not null, then put a "<br>" tag and then put the name of the file there.
(In the future, the name of the file will be a link to the file itself. But
this is not possible with the current implementation because uploaded files
are not stored inside the public directory.)