Feature request on the doctrine branch: 2795870	Doctrine Integration
Adding a test for the User model.

also had to fix the FismaUnitTest base class. It had overridden the setUp() method but had changed the signature, which breaks any test which doesn't call the setup manually. I modified the method back to its correct signature. If you want to load data fixtures in a model, then do it inside the test itself. 

(Although, loading data from the DB should not be necessary for most tests.)