Feature request on the doctrine branch: 2795870	Doctrine Integration
I made a mistake and was storing the current user in the registry, which is not persistent. I moved the current user into the Zend_Auth session namespace so that it will persist.

Also added a convenience function called User::currentUser() to get a reference to the current logged in user. I am working towards putting the dashboard back into functioning state.