Feature request on the doctrine branch: 2795870	Doctrine Integration
The dashboard is now working again, I think.

I realized that the usage of Zend_Acl is not exactly how it's intended to be used. If a resource is not defined, then Zend_Acl throws an exception, which I did not expect. So I added a quick hack to catch that exception. See the comments in Fisma_Acl.