Feature request on the doctrine branch: 2795870	Doctrine Integration
Adding some basic caching for the organization tree. This is only half of the formula. The cache needs to be invalidated whenever the summary counts change, which will need to be added as a listener in the finding model. Also, the cache needs to be invalidated once every day to account for new items which are overdue.