Feature request on the doctrine branch: 2795870	Doctrine Integration
Finally fixed the uploader on firefox. The issue was that the script was using DOM to reveal a hidden element. In safari, this works, but in firefox it causes the flash objects to get reloaded. The file uploader is written in flash, so it would get reloaded and forget what it had just done.

Next step is to test in IE.