Feature request on the trunk: #2795870 Doctrine Integration
Removing the generated models. We shouldn't version control generated code unless we have a good reason too.

This means that models *must* be generated before you can run OpenFISMA. But this will happen in the CI build script and will also happen in the installer (when the installer gets updated to work with doctrine.)