Feature request on the trunk: #2795870 Doctrine Integration
Fixing the doctrine-cli.php script. The "listener disabler" code was interfering with building the models. This required a new parameter so that the user listener function can operate in the context of a command line script. (Previously it was trying to reference the http request object, which does not work in CLI context)