Feature request on the trunk: #2795870 Doctrine Integration
Fixing the doctrine build issue. The problem is that Doctrine persists data fixtures in a weird order, and for some reason User gets persisted before Configuration, but User actually has a dependency on Configuration. (It needs to lookup the 'hash_type' configuration so that it can hash the user password.)

The quickest way to fix this is to modify Doctrine itself to sort the models alphabetically before persisting them. This will put Configuration before User, but it's just a band-aid. It may take some time to research a better solution.