Feature request on the trunk: #2795870 Doctrine Integration
When I first created the models, I left a lot of fields defined as "string" without specifying the length. I didn't realize that Doctrine will turn those into BLOB fields in the database.

The fix is to define them as string(255) -- unless there is a business reason to make them shorter -- which Doctrine will turn into VARCHAR fields in the database. This is better for general performance.

Also adding a few default document types.