Feature request on the trunk: #2795870 Doctrine Integration
I think I've got all of the finding notifications working now, but I see there is a really serious error with the mitigation strategy submit/approve/revise process. So I'm going to check in here so I can work on the two issues separately. But this may not be 100% correct yet.