Feature request on the trunk: #2795870 Doctrine Integration
I think i've fixed all of the ACL issues around modifying findings now. The big problem was that ACL for findings needs to take into account whether the user is allowed to access that system or not. Otherwise it would be pretty simple for a user to type in the path to a finding directly and bypass the simplistic security checks that the controller performs.

The main change in this commit is to make sure that the organization is checked all points, both in the view and in the finding listener, to make sure a user doens't see controls they shouldn't have, and also can't bypass the controls by posting data directly into openfisma.