Bug fix on the release branch: OFJ-25 Asset read privilege is not enforced
Modified the finding create form to be a little more readable. It's still less than desirable. The asset picker should be a popup instead of a group of controls. But until then, this will do.

It also hides asset information on the finding detail screen if the user is not allowed to see it.