Improvement on the release branch: OFJ-59 Add drag and drop to the Organization tree view
I noticed that Doctrine itself actually has a bug where it will calculate bogus values for the level field when you move subtrees around. I have reported this bug and a patch to the Doctrine project (See comments for OFJ-59) but until that patch is released, I need to fix this on our own copy.