Improvement on the release branch: OFJ-59 Add drag and drop to the Organization tree view
The select menus on the create finding page and finding inject page are now fixed.

The root cause was that the implementation for User::getOrganizations() was flawed and would not return the organizations in the proper, hierarchical order. I took this opportunity to refactor that method.

Now it calls a separate method, User::getOrganizationsQuery(), which return a Doctrine_Query object that can be used to get the user's organizations.