Bug fix on the release branch: OFJ-76 Email generates invalid links
This was tricky to fix. The host_url is initially defined in the Configuration.yml file. This setting needs to be written at install time. The installer is pretty messed up, so while I was in there I cleaned up the error handling a bit. The email template was also busted, not taking into account LDAP settings and including some random stuff at the end.