Bug fix on the release branch: OFJ-80 System search does not work
I'm approaching this bug with a major refactring. The current system for indexing is very limiting, and it hard codes lots of inconvient defaults.

I'm refactoring the listener to use Doctrine introspection instead. By defining search criteria in the YAML models, the new IndexListener will know how to create indexes automatically. This is far easier to maintain, and it also drastically reduces the amount of code required to implement this functionality.

This does require existing indexes to be destroyed and rebuilt, however. The 'create-indexes' script is now called 'optimize-indexes' to reflect more accurately what it's purpose is -- *but* it will still create missing indexes as it optimizes.