Bug fix on the release branch: OFJ-80 System search does not work
This is the 2nd step. I fixed a bugs from the first checkin. Then I added the ability to index fields in related models (its a little hacky). Then I updated some models (such as Asset) which would need to index foreign fields (e.g. if you want to search the asset table by the name of the product.

Everything is done except for the Finding itself. I started working on it, but I want to checkin now before I keep going. So this checkin is partially broken.