Bug fix on the release branch: OFJ-80 System search does not work
Finished the finding search and polished and tested. It's almost 100% repaired at this point. Good enough for tagging the next release, anyway.

I broke down and fixed a few other annoying bugs at the same time. This is bad form though -- we shouldn't do that.

The main thing missing right now is that if you change the value of a related record, the lucene index won't update for that related record. Example: change the name of a system and then search the findings for that system name. That's because you'd need to reindex every single finding associated with that system. Its far too much work to do during a page load, so it really needs to be a spawned process which runs in the background. But that requires a lot more architecture, and users won't search this way very often, so I'm closing this even with that caveat.