Bug fix on the release branch: OFJ-93 Short term fix for corrupted Zend Search Lucene indexes
Use try/catch with empty blocks to cover up errors in the listener (bad form, yes, but a necessary evil). Use try/catch in the finding search to display ZSL exceptions as user error messages (again, not great, since the ZSL error messages are not usually appropriate for end users).