Bug fix on the release branch: OFJ-109 Date parsing bug on finding creation
This is caused because the listener which sets the next action due date runs BEFORE the behavior which sets the createdAt timestamp.

The fix is to use Fisma::now() instead, which means that the next action due date may be a few milliseconds different from its true value. This would only be a problem if this happened right at midnight however. I'm willing to live with this compromise.

I don't think its possible to force the behavior to run first.