Improvement on the release branch: OFJ-121: REOPEN -Add the ability to leave a comment when approving a finding
Second step: format plain text as html (using a new view helper), re-style the page a bit so that larger amounts of text can be displayed as comments next to the approval.

I inserted an IE6 hack, because the new style relies on the first-child pseudo selector which IE6 doesn't work. So it won't look as pretty on IE6, but will still function.