Bug fix on the release branch: OFJ-52 The notification script doesn't send notifications
Finally, fix the original problem, which was an incorrect query.

Our attempts to do this with Doctrine_Query could not succeed. I resorted to using Doctrine_RawSql, which is not portable but at eleast hydrates real objects. So this is not ideal but it IS better than using the raw connection handle and native driver functions.