Bug fix on the release branch: OFJ-146 Create account failure when the password of new user is the same as current login account's
Move the password check from the validator into the listener. The validator doesn't have access to the object which is being modified, therefore it is not suitable for checking password history.

Also, to make this work, need to add a check to only generate 1 salt per user. That account keeps the same salt in the future so that we can compare hashed passwords.