Improvement on the release branch: OFJ-177 The finding controller summary-data action returns extraneous data
Remove some of the unnecessary data and then add the ability to gzip the response.

While working on this I noticed that ZF by default doesn't run your .json.phtml view scripts. So I enabled my .json.phtml view script in this controller, but this is something that needs a redesign.

Ideally we would extend or modify the context switch helper to provide automatic gzipping of json data (if supported by the client) and automatic debug views. (Currently you have to manually make a debug view for each json view.)