Bug on release branch: OFJ-264 Finding workflow is broken
My first step in fixing this is to refactor Finding Listener. Its functionality has been moved into the Finding model itself. Some of that code stays inside the hooks for now, but I moved as much as possible into separate mutator functions which is cleaner. 

Its also easier to debug and test. I'm checking in a few tests for the finding model, too.

It's still broken, but I'm going to take a break on this ticket for a bit.