Bug fix on the release branch OFJ-171 The system occurs errors when some parameters on the configuration page are set as an invalid value
Renamed some of the validators, so that the names reflect what the validators actually are.
Added additional validators to fields.
Modified Hostname validator to use the ALLOW_ALL option, so that all valid hostnames/IPs/local domains can be used as the hostname.
Modified some of the english on the forms, where it didn't make sense.