Bug fix on the release branch
OFJ-135 The statistic of finding status is incorrect on dashboard screen after one finding closed.

Check to make sure that the id/nickname in CurrentEvaluation join are not null, before setting the results array to 
count them. The count is 0, since the joined fields are null. This is basically extraneous data due to doing a leftJoin 
and not having any data on the currentevaluation join.