Bug fix on the relase branch: OFJ-288 Design of authentication is flawed
This was a considerable effort. I refactored as explained in the documentation submitted previously. I wanted to write tests for this but the whole module -- including User and AuthController -- are just way too coupled with logging, notificaitons, etc. to do any testing.

Not 100% tested, but I did manually test most of the execution paths, such as various account locks, incorrect passwords, LDAP or database, etc. Going to do a bit more testing before officially submitting.