Bug fix on the release branch
OFJ-321 Installation fails

Change the constructor on the User model to only use Configuration model if it's available. If it's not available, we'll default to sha1 for the object being constructed.
Modify the User fixture so that the hashType is defined by default to sha1.
Modify the Install controller so that once the root user is loaded by the fixture, the hashType is set to whatever the user has defined during installation.