Bug fix on the trunk
OFJ-4 Improve 508 Compliance

Address Standard A: Text Equivalents.
Change embed to object, and add in some text advising that Flash is required to view the chart.
It would be nice to eventually generate normal images to put in the place of the Flash charts, but this will bring us into compliance for now.