Improvement on the trunk OFJ-268 Refactor listeners into mutators/accessors
Tackling System next. I added some tests for the functionality which was merged into the System model, but I know I broke something because I some of the functionality in the listener didn't make any sense at all. 

I also removed the ugly MapValues stuff which we were not using correctly, but I haven't figured out what the collateral effects of that are yet.