Bug fix on trunk
OFJ-152 Asset form is incorrect

This is the first of a few steps to correct the Asset form. We'll also end up with support across the application for AutoComplete fields in forms once this task is complete. Committing this now so it doesn't get lost.

Create a Fisma JS namespace
Create directory structure for Fisma specific JS
Create Fisma.AutoComplete namespace