Improvement on the trunk: OFJ-339 Decouple Configuration Class
Add a configuration interface and two implementations: one is the existing implementation based on the doctrine table, the other is a test harness implementation based on array storage.

Add static getter/setter on the Fisma class to manipulate configuration objects.

Update all references to system configuration to use this new object.