Improvement on the trunk: OFJ-354 Add audit log behavior
Adding the basic behavior. This behavior can be attached to a model with 'actAs' and will generate a corresponding audit log table for each model which enables this behavior.

I had to patch Doctrine_Import_Builder to suppress warnings because while it supports loading behaviors out of a non-standard location, it won't do so without emitting warnings.

I haven't added any of the actual audit log logic yet. That will come in the next commit.