Improvement on the trunk: OFJ-354 Add audit log behavior
This is most of the core behavior. I have not implemented the fetch() or query() methods yet, but I think those are going to be simple. I've updated the Asset model to use this behavior as an example, but the the audit log data is not displayed anywhere in the interface.

My next step is to add this behavior to the User and Finding models (which are the only ones which have audit logs currently) and write migrations to move the existing audit logs into the new tables.