Improvement on the trunk: OFJ-354 Add audit log behavior
Add audit logging for the user model, including updated YAML, migrations to copy data from the existing account_log table, and updating the auth controller, user controller, and user model to use this new log mechanism. Removed all old log mechanisms and exposed the new one with a link on the "view user" page. 

This also required some tweaks to the design. Logging can now be disabled on a class-by-class basis. I also realized I hhadn't finished the implementation for getAuditLog()->fetch() and getAuditLog()->query(), so those are in there now.

Next step is to convert the finding model.