Improvement on the trunk: OFJ-366 Implement a Doctrine_Record subclass that has a getOriginalValues() method
Add a new base class called Fisma_Record. This extends Doctrine_Record and provides a method for getting the "original value" of a field. See the ticket for an explanation of how this differs from the getLastModified() method.

I am also updating the Fisma class to configure Doctrine to use Fisma_Record as the base class when auto-generating model classes.

This commit is implementation only. The next commit will contain a test.