Improvement on the trunk: OFJ-354 Add audit log behavior
Add the behavior for Finding. This required a few more tweaks and had a few blocking issues for the audit loggable behavior. At this point I think the behavior is stable and ready for code review.

Also I removed a few classes related to the legacy user model audit logging.