Improvement on the trunk: OFJ-382 Add a "sample user" fixture
This turned out to be far more complicated than I had anticipated. There is no easy way in Doctrine to have sample data that has relations to standard fixtures.

So I went a bit out of my way and modified the doctrine-cli.php script to accept a switch called 'sample-data' which will perform a mini-build by merging standard fixtures and sample fixtures into one directory and then running the Doctrine CLI.

In the process I found that I need to use a suffix on all of the Organization YAML objects in order to prevent namespace collisions. This is now building pretty smoothly.