Bug on the trunk: OFJ-388 Build is broken because doctrine-cli.php tries to disable listeners before models have been generated
Fixed as described on the ticket. I created a new subclass for Doctrine_Record_Listener which has a static method for disabling it. Each listener which we might want to disable needs to subclass this class and check whether it is enabled before it executes any of the lifecycle hooks, such as preSave().

This should fix the build.