New feature on trunk
OFJ-290 Create/update nessus plugin

Only duplicate asset/product handling in the same scan remains to be completed.
AppDetective not tested yet.
Refactored to be much faster, moving all blocking DB operations (writes) to the end, all DB functions wrapped in a transaction.
Only MODERATE and HIGH Nessus findings are imported, as most LOW findings are pretty much useless and just informational.
The RNG file needs to be fixed, as at the moment, it's only validating against the nessus report that is attached to the JIRA issue.