Bug on the trunk: OFJ-343 There is something wrong in the table on finding summary page
Renamed TreeTable.js to FindingSummary.js (and same for the CSS file) and refactored to look like other javascript in the Fisma namespace. Also replaced the innerHTML stuff with DOM method.

This is just a cleanup step for debugging the actual bug, which may require more refactoring since this is a pretty hairy module right now.