Improvement on the trunk: OFJ-416 When creating a new system/organization, the current user should have access to it by default
This is trivial to do for Organization, but a little trickier for System. On the system controller we have a method called saveValue which is a good place to hook into, but this method is called for updates and inserts, and Doctrine's Unit of Work doesn't let us add our relation when we first create the object.

So I had to use a flag and set up the relation after all the intermediary objects are already persisted. Its a little ugly but it works. We can clean this up when we redesign the BaseController.