Bug on the trunk: OFJ-417 CRUD Notifications are not filtered by the underlying organization/user ACL
This is the massive first step. Refactor the ACL and update all ACL references to use this new design.

I went through all functionality in the app as the sample user, so nothing is horribly broken, but there still might be some small flaws with various permutations of privileges.

We will devote some extra QA to this topic over the next week.