Bug on the trunk OFJ-232 Unique key violations result in undecipherable system errors for users
Fix an issue with the doctrine build script. It was failing because the root user has an email address 'sample@agency.gov', which fails the email validation because it cannot resolve 'agency.gov'.

I changed two things to work around this:
1) Set the email validator options check_mx=false, which disables the DNS MX lookup. 
2) Change root's default e-mail address to openfisma-default-install@googlegroups.com. This is a write-only group (unless you are a member) so we won't have hundreds of people bouncing emails from 'agency.gov'. We can also check the group to verify that e-mail is working.

Also, I updated the Fisma_Record error stack method to provide a better error message when an email validation fails.