Improvement on the trunk OFJ-360 Remove doctrine hack from User.php
The doctrine-cli script was broken due to the previous commit on this issue. This one is a little trickier than the installer. I created a temporary configuration array to use while rebuilding the db, and then after the rebuild finishes I switch it to a database-backed configuration object.

This repairs the unit tests and should fix the broken build.