Bug on the trunk OFJ-434 Password emails send the hash instead of actual password
Use mapValue as described in the issue comments. This works pretty well.

However I noticed while testing it that my fix in r2899 broke the email send, because I was saving the User object before it had any values in it. I moved the save() down below the merge() to remedy this.

I tested out the email and it does send the plain text password now.