Bug on the trunk OFJ-433 An error occurs when upload scan results
Update app detective plugin to omit the name, vendor, and version when mapping product information; it maps only the CPE identifier or else NULL.

Update abstract base class to not match duplicates on name, vendor, or version if the CPE ID is available, since the CPE ID is a more reliable key. If the CPE ID is null, then matching is performed on the name, vendor, and version instead.