Bug on the trunk OFJ-453 Hide the button "discard changes" and "save changes" for read-only user on finding detail page
Modify Fisma_Acl to support wildcard queries on the privilege name. This is backwards compatible with the previous Fisma_Acl, so no changes to the app are required.

Update the finding view script to use the wildcard ACL to see if a user has any update_* privileges on the current finding object.