Improvement on trunk
OFJ-8 Performance Engineering

Cache ACL object

Modify invalidateAcl() to also remove the ACL record from the cache.
Add postSave() hooks to invalidate the ACL of affected users when role privileges are modified, member roles are changed, or member organizations are modified. 
With this change, it should be possible for an administrator to change a user's privileges/organizations/roles, and for the change to take affect immediately, even if the user is currently logged in.