Bug fix on branch
OFJ-477 Remove the "Edit" link on organization tree page when the user has no edit rights

Removing the edit link entirely. The way the JS is built, it would be a huge LOE to assign context menus per item. View link is still there, and edit is accessible from view, if you have the privilege.