Improvement on the trunk: OFJ-509 Convert IR interface to tabbed interface
Create a Fisma_Yui_TabView class to wrap the tabview code. The actual view is rendered in a view script to keep the class neat. I migrated the Finding view page over to use this as a proof-of-concept.

I'm also re-adding the optimize-indexes script which for some reason was deleted during the IR-to-trunk merge but was needed during this task.