New feature on trunk OFJ-372 Add audit logging for all incidents
As explained in Jira, I accidentally merged a bit of this task with a prior commit (r3050). So adding the audit log behavior isn't seen in this commit.

This commit includes the necessary actions to display the audit log, plus a few custom log entries for lifecycle events for the incident (reject, close, etc.). I also updated some metadata in the schema file to make the logical names look prettier (turns out you don't escape quotes in YAML).

Finally, update the log behavior to sort on ID as a secondary key. This means that log events that happen 'simultaneously' will be output in the same order they were put it in. This ensures a consistent ordering for these events, regardless of the physical layout on disk.